partly applied code review changes
